my_app/requirements.txt
pyaml==20.4.0
PyYAML==5.3.1